refactor: 주차장 머지 충돌 해결 -  #47
